# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple API to create and manage wishlists with any sort of products or content",
    "title": "Go Wishlist API",
    "contact": {
      "name": "Stas Bukovskyi",
      "email": "stas.bukovskyi@gmail.com"
    },
    "version": "1.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "paths": {
...
    "/api/v1/wishlist-items/": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Create a new wishlist item",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "wishlist-items"
        ],
        "summary": "Create a new wishlist item",
        "parameters": [
          {
            "description": "create wishlist item request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.AddWishlistItemRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "wishlist item",
            "schema": {
              "$ref": "#/definitions/entity.WishlistItem"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          }
        }
      }
    },
    "/api/v1/wishlist-items/{id}": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Get wishlist item by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "wishlist-items"
        ],
        "summary": "Get wishlist item",
        "parameters": [
          {
            "type": "string",
            "description": "wishlist item id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "wishlist item",
            "schema": {
              "$ref": "#/definitions/entity.WishlistItem"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Create existing wishlist item by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "wishlist-items"
        ],
        "summary": "Update a wishlist item",
        "parameters": [
          {
            "type": "string",
            "description": "wishlist item id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "update wishlist item request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateWishlistItemRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "wishlist item",
            "schema": {
              "$ref": "#/definitions/entity.WishlistItem"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Delete wishlist item by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "wishlist-items"
        ],
        "summary": "Delete wishlist item",
        "parameters": [
          {
            "type": "string",
            "description": "wishlist item id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "wishlist item",
            "schema": {
              "$ref": "#/definitions/entity.WishlistItem"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/httperrs.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "entity.Image": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "image_name": {
          "type": "string"
        },
        "updated_at": {
          "type": "integer"
        },
        "url": {
          "type": "string"
        },
        "wishlist_item_id": {
          "type": "string"
        }
      }
    },
...
    "entity.WishlistItem": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/entity.Image"
          }
        },
        "price": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "updated_at": {
          "type": "integer"
        },
        "url": {
          "type": "string"
        },
        "wishlist_id": {
          "type": "string"
        }
      }
    },
    "httperrs.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/httperrs.ServiceError"
        }
      }
    },
    "httperrs.ServiceError": {
      "type": "object",
      "properties": {
        "kind": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "param": {
          "type": "string"
        }
      }
    },
    "request.AddWishlistItemRequest": {
      "type": "object",
      "required": [
        "url",
        "wishlist_id"
      ],
      "properties": {
        "description": {
          "type": "string"
        },
        "image_ids": {
          "type": "array",
          "maxItems": 5,
          "items": {
            "type": "string"
          }
        },
        "price": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "wishlist_id": {
          "type": "string"
        }
      }
    },
...
    "request.UpdateWishlistItemRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "image_ids": {
          "type": "array",
          "maxItems": 5,
          "items": {
            "type": "string"
          }
        },
        "price": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}